requirements.txt
amqp==2.4.2
arrow==0.13.1
attrs==19.1.0
Babel==2.6.0
backcall==0.1.0
beautifulsoup4==4.7.1
billiard==3.5.0.5
bleach==3.1.0
celery==4.2.2
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
cssselect==1.0.3
cycler==0.10.0
decorator==4.4.0
defusedxml==0.5.0
Django==2.1.7
django-cors-headers==2.5.2
django-rest-framework-rules==1.0.0
djangorestframework==3.9.2
djangorestframework-simplejwt==4.3.0
djangorestframework-stubs==0.4.2
dnspython==1.16.0
entrypoints==0.3
eventlet==0.24.1
flower==0.9.3
greenlet==0.4.15
idna==2.8
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterthemes==0.20.0
kiwisolver==1.1.0
kombu==4.3.0
lesscpy==0.13.0
lxml==4.3.3
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
monotonic==1.5
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.7
numpy==1.16.3
pandocfilters==1.4.2
parso==0.3.4
pendulum==2.0.4
pickleshare==0.7.5
ply==3.11
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psutil==5.6.1
Pygments==2.3.1
PyJWT==1.7.1
pyparsing==2.4.0
pyquery==1.4.0
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2018.9
pytzdata==2018.9
pywinpty==0.5.5
pyzmq==18.0.1
qtconsole==4.4.3
redis==3.2.1
requests==2.21.0
rules==2.0.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.9
terminado==0.8.2
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.1
vine==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
